+++ playground/playground.js
@@ -3,16 +3,6 @@
 
 'use strict';
 
-/*
-
-  System A
-
-  Input(type: Type, opts(A): maybe(Obj))
-  input.render() -> A
-  input.getValue(depth: maybe(Num)) -> validate.Result | type
-
-*/
-
 var React = require('react');
 var cx =    require('react/lib/cx');
 var t =     require('tcomb-validation');
@@ -26,7 +16,6 @@ var Bool =        t.Bool;
 var Obj =         t.Obj;
 var Arr =         t.Arr;
 var Func =        t.Func;
-var irriducible = t.irriducible;
 var subtype =     t.subtype;
 var maybe =       t.maybe;
 var enums =       t.enums;
@@ -35,12 +24,10 @@ var struct =      t.struct;
 var tuple =       t.tuple;
 var func =        t.func;
 var mixin =       t.util.mixin;
-var isType =      t.util.isType;
+var Type =        t.Type;
 var getKind =     t.util.getKind;
 var getName =     t.util.getName;
 var Result =      t.validate.Result;
-
-var Type = irriducible('Type', isType);
 
 var Order = enums({
   asc: function (a, b) {
@@ -79,12 +66,11 @@ function humanize(s){
   return capitalize(underscored(s).replace(/_id$/,'').replace(/_/g, ' '));
 }
 
-function stripMaybeOrSubtype(type) {
-  var kind = getKind(type);
-  if (kind === 'maybe' || kind === 'subtype') {
-    return stripMaybeOrSubtype(type.meta.type);
-  }
-  return type;
+function stripContainerType(type, types) {
+  types = types || {maybe: 1, subtype: 1};
+  return types.hasOwnProperty(getKind(type)) ?
+    stripContainerType(type.meta.type, types) :
+    type;
 }
 
 function getOrElse(value, defaultValue) {
@@ -245,8 +231,8 @@ function moveDown(arr, i) {
 // default React class methods
 //
 
-function getInitialState() {
-  return { hasError: false };
+function getInitialState(hasError) {
+  return { hasError: getOrElse(hasError, false) };
 }
 
 function getValue(type, rawValue) {
@@ -264,7 +250,7 @@ function getValue(type, rawValue) {
 // ========================
 
 function getInput(type) {
-  type = stripMaybeOrSubtype(type);
+  type = stripContainerType(type);
   var kind = getKind(type);
   var ret = options.inputs[kind];
   if (Func.is(ret)) {
@@ -302,13 +288,15 @@ function textboxOpts(type) {
     breakpoints:  maybe(Breakpoints),
     height:       maybe(Size),
     onKeyDown:    maybe(Func),
-    onChange:     maybe(Func)
+    onChange:     maybe(Func),
+    message:      maybe(Str),
+    hasError:     maybe(Bool)
   }, 'TextboxOpts');
 }
 
 function textbox(type, opts) {
 
-  assert(isType(type));
+  assert(Type.is(type));
 
   opts = new (textboxOpts(type))(opts || {});
 
@@ -410,18 +398,10 @@ function textbox(type, opts) {
 // select
 //
 
-// select accepts only enums or maybe(enums)
-var EnumType = subtype(Type, function (type) {
-  var kind = getKind(type);
-  if (kind === 'enums') {
-    return true;
-  }
-  return kind === 'maybe' && getKind(type.meta.type) === 'enums';
-}, 'EnumType');
-
 function selectOpts(type) {
   return struct({
     ctx:          Any,
+    name:         maybe(Str),
     options:      Any,
     value:        maybe(type),
     label:        Any,
@@ -431,21 +411,25 @@ function selectOpts(type) {
     order:        maybe(Order),
     disabled:     maybe(Bool),
     breakpoints:  maybe(Breakpoints),
-    height:       maybe(Size)
+    height:       maybe(Size),
+    multiple:     maybe(Bool)
   }, 'SelectOpts');
 }
 
 function select(type, opts) {
 
-  type = EnumType(type);
+  assert(Type.is(type));
+
   opts = new (selectOpts(type))(opts || {});
 
-  var Enum = stripMaybeOrSubtype(type);
-  var emptyValue = opts.emptyOption ? opts.emptyOption.value : null;
+  var Enum = stripContainerType(type, {maybe: 1, subtype: 1, list: 1});
+  var isMultiple = opts.multiple === true;
+  var emptyOption = isMultiple ? null : opts.emptyOption;
+  var emptyValue = emptyOption ? emptyOption.value : null;
   var defaultValue = getOrElse(opts.value, emptyValue);
   var label = getLabel(opts.label, opts.breakpoints);
   var help = getHelp(opts.help);
-  var options = getOptions(opts.options || Enum, opts.order, opts.emptyOption);
+  var options = getOptions(opts.options || Enum, opts.order, emptyOption);
 
   var inputClasses = {
     'form-control': true
@@ -461,8 +445,18 @@ function select(type, opts) {
     getInitialState: getInitialState,
     
     getRawValue: function () {
-      var value = this.refs.input.getDOMNode().value;
-      return value === emptyValue ? null : value;
+      var select = this.refs.input.getDOMNode();
+      if (isMultiple) {
+        var values = [];
+        for (var i = 0, len = select.options.length ; i < len ; i++ ) {
+            var option = select.options[i];
+            if (option.selected) {
+              values.push(option.value);
+            }
+        }
+        return values;
+      }
+      return select.value === emptyValue ? null : select.value;
     },
     
     getValue: getValue(type),
@@ -477,10 +471,12 @@ function select(type, opts) {
       var input = (
         React.DOM.select({
           ref: "input", 
+          name: opts.name, 
           className: cx(inputClasses), 
           disabled: opts.disabled, 
           readOnly: opts.readOnly, 
-          defaultValue: defaultValue}, 
+          defaultValue: defaultValue, 
+          multiple: isMultiple}, 
           options
         )
       );
@@ -513,6 +509,7 @@ function select(type, opts) {
 function radioOpts(type) {
   return struct({
     ctx:          Any,
+    name:         maybe(Str),
     value:        Any,
     label:        Any,
     help:         Any, 
@@ -524,16 +521,17 @@ function radioOpts(type) {
 
 function radio(type, opts) {
 
-  type = EnumType(type);
+  assert(Type.is(type));
+
   opts = new (radioOpts(type))(opts || {});
 
-  var Enum = stripMaybeOrSubtype(type);
+  var Enum = stripContainerType(type);
   var defaultValue = getOrElse(opts.value, null);
   var label = getLabel(opts.label, opts.breakpoints);
   var help = getHelp(opts.help);
   var choices = getChoices(Enum.meta.map, opts.order);
   var len = choices.length;
-  var name = uuid();
+  var name = opts.name || uuid();
 
   return React.createClass({
     
@@ -598,17 +596,10 @@ function radio(type, opts) {
 // checkbox
 //
 
-// checkbox accepts only Bool or subtypes of Bool
-var CheckboxType = subtype(Type, function (type) {
-  if (type === Bool) {
-    return true;
-  }
-  return getKind(type) === 'subtype' && type.meta.type === Bool;
-}, 'CheckboxType');
-
 function checkboxOpts(type) {
   return struct({
     ctx:          Any,
+    name:         maybe(Str),
     value:        maybe(type),
     label:        Any,
     help:         Any, 
@@ -619,7 +610,8 @@ function checkboxOpts(type) {
 
 function checkbox(type, opts) {
 
-  type = CheckboxType(type);
+  assert(Type.is(type));
+
   opts = new (checkboxOpts(type))(opts || {});
 
   var defaultValue = getOrElse(opts.value, false);
@@ -647,7 +639,7 @@ function checkbox(type, opts) {
       var input = (
         React.DOM.div({className: "checkbox"}, 
           React.DOM.label(null, 
-            React.DOM.input({ref: "input", type: "checkbox", defaultChecked: defaultValue}), " ", opts.label
+            React.DOM.input({ref: "input", type: "checkbox", name: opts.name, defaultChecked: defaultValue}), " ", opts.label
           )
         )
       );
@@ -676,15 +668,6 @@ function checkbox(type, opts) {
 // forms
 //
 
-// createForm accepts only structs or subtypes of a struct
-var FormType = subtype(Type, function (type) {
-  var kind = getKind(type)
-  if (kind === 'struct') {
-    return true;
-  }
-  return kind === 'subtype' && getKind(type.meta.type) === 'struct';
-}, 'FormType');
-
 var FormAuto = enums.of('none placeholders labels', 'FormAuto');
 
 var FormOpts = struct({
@@ -700,10 +683,11 @@ var FormOpts = struct({
 
 function createForm(type, opts) {
 
-  type = FormType(type);
+  assert(Type.is(type));
+
   opts = new FormOpts(opts || {});
 
-  var Struct = stripMaybeOrSubtype(type);
+  var Struct = stripContainerType(type);
   var props = Struct.meta.props;
   var keys = Object.keys(props);
   var order = opts.order || keys;
@@ -729,7 +713,7 @@ function createForm(type, opts) {
     var Input = o.input ? o.input : getInput(type);
 
     // handle optional fields auto label
-    var optional = getKind(type) === 'maybe' ? options.optionalText : '';
+    var optional = getKind(type) === 'maybe' ? options.bundle.optional : '';
 
     // lists, forms, checkboxes and radios must always have a label
     if (Input === createList || Input === createForm || Input === checkbox || Input === radio) {
@@ -747,7 +731,7 @@ function createForm(type, opts) {
         }
       } else if (auto === 'placeholders' && !o.label) {
         if (Input === select) {
-          o.emptyOption = o.emptyOption || {value: '', text: humanize('Select your ' + name + optional)};
+          o.emptyOption = o.emptyOption || {value: '', text: humanize(options.bundle.select + name + optional)};
         } else if (Input === textbox) {
           o.placeholder = o.placeholder || humanize(name + optional);
         }
@@ -803,7 +787,7 @@ function createForm(type, opts) {
       });
 
       return (
-        React.DOM.div({className: cx(classes)}, 
+        React.DOM.fieldset({className: cx(classes)}, 
           label, 
           children
         )
@@ -818,15 +802,6 @@ function createForm(type, opts) {
 // lists
 //
 
-// createList accepts only lists or subtypes of a lists
-var ListType = subtype(Type, function (type) {
-  var kind = getKind(type)
-  if (kind === 'list') {
-    return true;
-  }
-  return kind === 'subtype' && getKind(type.meta.type) === 'list';
-}, 'ListType');
-
 var ListOpts = struct({
   ctx:            Any,
   value:          maybe(Arr),
@@ -840,11 +815,12 @@ var ListOpts = struct({
 
 function createList(type, opts) {
 
-  type = ListType(type);
+  assert(Type.is(type));
+
   opts = new ListOpts(opts || {});
 
-  var List = stripMaybeOrSubtype(type);
-  var ItemType = stripMaybeOrSubtype(List.meta.type);
+  var List = type;
+  var ItemType = stripContainerType(List.meta.type);
   var Input = opts.input || getInput(ItemType);
   var defaultValue = getOrElse(opts.value, []);
   var label = getLabel(opts.label);
@@ -945,9 +921,9 @@ function createList(type, opts) {
             ), 
             React.DOM.div({className: "col-md-5"}, 
               React.DOM.div({className: "btn-group"}, 
-                opts.disableRemove ? null : React.DOM.button({className: "btn btn-default btn-remove", onClick: this.remove.bind(this, i)}, "Remove"), 
-                !opts.disableOrder ? React.DOM.button({className: "btn btn-default btn-move-up", onClick: this.moveUp.bind(this, i)}, "Up") : null, 
-                !opts.disableOrder ? React.DOM.button({className: "btn btn-default btn-move-down", onClick: this.moveDown.bind(this, i)}, "Down") : null
+                opts.disableRemove ? null : React.DOM.button({className: "btn btn-default btn-remove", onClick: this.remove.bind(this, i)}, options.bundle.remove), 
+                !opts.disableOrder ? React.DOM.button({className: "btn btn-default btn-move-up", onClick: this.moveUp.bind(this, i)}, options.bundle.up) : null, 
+                !opts.disableOrder ? React.DOM.button({className: "btn btn-default btn-move-down", onClick: this.moveDown.bind(this, i)}, options.bundle.down) : null
               )
             )
           )
@@ -956,12 +932,12 @@ function createList(type, opts) {
 
       var btnAdd = opts.disableAdd ? null : (
         React.DOM.div({className: "form-group"}, 
-          React.DOM.button({className: "btn btn-default btn-add", onClick: this.add}, "Add")
+          React.DOM.button({className: "btn btn-default btn-add", onClick: this.add}, options.bundle.add)
         )
       );
 
       return (
-        React.DOM.div({className: cx(classes)}, 
+        React.DOM.fieldset({className: cx(classes)}, 
           label, 
           children, 
           btnAdd
@@ -974,8 +950,9 @@ function createList(type, opts) {
 }
 
 function create(type, opts) {
-  var kind = getKind(stripMaybeOrSubtype(type));
-  return kind === 'struct' ? createForm(type, opts) : createList(type, opts);
+  return getKind(stripContainerType(type)) === 'struct' ? 
+    createForm(type, opts) : 
+    createList(type, opts);
 }
 
 // ===============================
@@ -983,7 +960,14 @@ function create(type, opts) {
 // ===============================
 
 var options = {
-  optionalText: ' (optional)',
+  bundle: {
+    select: 'Select your ',
+    optional: ' (optional)',
+    add: 'Add',
+    remove: 'Remove',
+    up: 'Up',
+    down: 'Down'
+  },
   inputs: {
     irriducible: {
       Bool: checkbox
@@ -22671,7 +22655,7 @@ module.exports = require('./lib/React');
   var maybe = t.maybe;
   var list = t.list;
 
-  var isType = t.util.isType;
+  var Type = t.Type;
   var assert = t.assert;
   var getName = t.util.getName;
   var mixin = t.util.mixin;
@@ -22736,9 +22720,7 @@ module.exports = require('./lib/React');
   }
 
   //
-  // validation functions
-  // one for each kind expect for `any`, `primitive`, `enums` 
-  // which are handled the same way
+  // validation functions, one for each kind
   //
 
   function validateIrriducible(value, type, opts) {
@@ -22784,7 +22766,7 @@ module.exports = require('./lib/React');
   }
 
   function validateMaybe(value, type, opts) {
-    assert(isType(type) && type.meta.kind === 'maybe');
+    assert(Type.is(type) && type.meta.kind === 'maybe');
 
     if (!Nil.is(value)) {
       return _validate(value, type.meta.type, opts);
@@ -22892,7 +22874,14 @@ module.exports = require('./lib/React');
     var errors = [];
     for (var k in value) {
       if (value.hasOwnProperty(k)) {
-        var result = _validate(value[k], type.meta.type, {path: opts.path.concat([k]), messages: getMessage(opts.messages, ':type')});
+        // domain
+        var result = _validate(k, type.meta.domain, {path: opts.path.concat([k]), messages: getMessage(opts.messages, ':domain')});
+        if (!result.isValid()) {
+          isValid = false;
+          errors = errors.concat(result.errors);
+        }        
+        // codomain
+        result = _validate(value[k], type.meta.codomain, {path: opts.path.concat([k]), messages: getMessage(opts.messages, ':codomain')});
         if (!result.isValid()) {
           isValid = false;
           errors = errors.concat(result.errors);
@@ -22934,7 +22923,7 @@ module.exports = require('./lib/React');
 
   function validate(value, type, opts) {
     opts = opts || {};
-    assert(isType(type), 'Invalid argument `type` of value `%j` supplied to `validate`, expected a type', type);
+    assert(Type.is(type), 'Invalid argument `type` of value `%j` supplied to `validate`, expected a type', type);
     assert(maybe(Arr).is(opts.path), 'Invalid argument `opts.path` of value `%j` supplied to `validate`, expected an `Arr`', opts.path);
 
     opts.path = opts.path || [];
@@ -22981,10 +22970,9 @@ module.exports = require('./lib/React');
     failed = true;
     throw new Error(message);
   }
-  
+
   var options = {
-    onFail: onFail,
-    update: null
+    onFail: onFail
   };
 
   function fail(message) {
@@ -23031,7 +23019,7 @@ module.exports = require('./lib/React');
   }
 
   function merge() {
-    return Array.prototype.reduce.call(arguments, function (x, y) {
+    return Array.prototype.reduce.call(arguments, function reducer(x, y) {
       return mixin(x, y, true);
     }, {});
   }
@@ -23069,44 +23057,86 @@ module.exports = require('./lib/React');
     j: function (x) { return JSON.stringify(x, replacer); }
   };
   
-  function isType(type) {
-    return Func.is(type) && Obj.is(type.meta);
-  }
-  
   function getName(type) {
-    assert(isType(type), 'Invalid argument `type` of value `%j` supplied to `getName()`, expected a type.', type);
+    assert(Type.is(type), 'Invalid argument `type` of value `%j` supplied to `getName()`, expected a type.', type);
     return type.meta.name;
   }
 
-  function deprecated(message) {
-    if (console && console.warn) {
-      console.warn(message);
-    }
-  }
-
   function getKind(type) {
-    assert(isType(type), 'Invalid argument `type` of value `%j` supplied to `geKind()`, expected a type.', type);
+    assert(Type.is(type), 'Invalid argument `type` of value `%j` supplied to `geKind()`, expected a type.', type);
     return type.meta.kind;
   }
 
-  function isKind(type, kind) {
-    deprecated('`isKind(type, kind)` is deprecated, use `getKind(type) === kind` instead');
-    return getKind(type) === kind;
-  }
-
   function blockNew(x, type) {
-    // DEBUG HINT: since in tcomb the only real constructors are those provided
-    // by `struct`, the `new` operator is forbidden for all types
     assert(!(x instanceof type), 'Operator `new` is forbidden for `%s`', getName(type));
   }
-  
-  function update() {
-    assert(Func.is(options.update), 'Missing `options.update` implementation');
-    /*jshint validthis:true*/
-    var T = this;
-    var value = options.update.apply(T, arguments);
-    return T(value);
+
+  function shallowCopy(x) {
+    return Arr.is(x) ? x.concat() : Obj.is(x) ? mixin({}, x) : x;
   }
+
+  function update(instance, spec) {
+    assert(!Nil.is(instance));
+    assert(Obj.is(spec));
+    var value = shallowCopy(instance);
+    for (var k in spec) {
+      if (spec.hasOwnProperty(k)) {
+        if (update.commands.hasOwnProperty(k)) {
+          assert(Object.keys(spec).length === 1);
+          return update.commands[k](spec[k], value);
+        } else {
+          value[k] = update(value[k], spec[k]);
+        }
+      }
+    }
+    return value;
+  }
+
+  update.commands = {
+    '$apply': function (f, value) {
+      assert(Func.is(f));
+      return f(value);
+    },
+    '$push': function (elements, arr) {
+      assert(Arr.is(elements));
+      assert(Arr.is(arr));
+      return arr.concat(elements);
+    },
+    '$remove': function (keys, obj) {
+      assert(Arr.is(keys));
+      assert(Obj.is(obj));
+      for (var i = 0, len = keys.length ; i < len ; i++ ) {
+        delete obj[keys[i]];
+      }
+      return obj;
+    },
+    '$set': function (value) {
+      return value;
+    },
+    '$splice': function (splices, arr) {
+      assert(list(Arr).is(splices));
+      assert(Arr.is(arr));
+      return splices.reduce(function (acc, splice) {
+        acc.splice.apply(acc, splice);
+        return acc;
+      }, arr);
+    },
+    '$swap': function (config, arr) {
+      assert(Obj.is(config));
+      assert(Num.is(config.from));
+      assert(Num.is(config.to));
+      assert(Arr.is(arr));
+      var element = arr[config.to];
+      arr[config.to] = arr[config.from];
+      arr[config.from] = element;
+      return arr;
+    },
+    '$unshift': function (elements, arr) {
+      assert(Arr.is(elements));
+      assert(Arr.is(arr));
+      return elements.concat(arr);
+    }
+  };
 
   //
   // irriducibles
@@ -23142,57 +23172,59 @@ module.exports = require('./lib/React');
     return Irriducible;
   }
 
-  var Any = irriducible('Any', function () {
+  var Any = irriducible('Any', function isAny() {
     return true;
   });
   
-  var Nil = irriducible('Nil', function (x) {
-    return x === null || x === undefined;
+  var Nil = irriducible('Nil', function isNil(x) {
+    return x === null || x === void 0;
   });
   
-  var Str = irriducible('Str', function (x) {
+  var Str = irriducible('Str', function isStr(x) {
     return typeof x === 'string';
   });
   
-  var Num = irriducible('Num', function (x) {
+  var Num = irriducible('Num', function isNum(x) {
     return typeof x === 'number' && isFinite(x) && !isNaN(x);
   });
   
-  var Bool = irriducible('Bool', function (x) {
+  var Bool = irriducible('Bool', function isBool(x) {
     return x === true || x === false;
   });
   
-  var Arr = irriducible('Arr', function (x) {
+  var Arr = irriducible('Arr', function isArr(x) {
     return x instanceof Array;
   });
   
-  var Obj = irriducible('Obj', function (x) {
+  var Obj = irriducible('Obj', function isObj(x) {
     return !Nil.is(x) && typeof x === 'object' && !Arr.is(x);
   });
   
-  var Func = irriducible('Func', function (x) {
+  var Func = irriducible('Func', function isFunc(x) {
     return typeof x === 'function';
   });
   
-  var Err = irriducible('Err', function (x) {
+  var Err = irriducible('Err', function isErr(x) {
     return x instanceof Error;
   });
   
-  var Re = irriducible('Re', function (x) {
+  var Re = irriducible('Re', function isRe(x) {
     return x instanceof RegExp;
   });
   
-  var Dat = irriducible('Dat', function (x) {
+  var Dat = irriducible('Dat', function isDat(x) {
     return x instanceof Date;
   });
 
-  var Type = irriducible('Type', isType);
+  var Type = irriducible('Type', function isType(x) {
+    return Func.is(x) && Obj.is(x.meta);
+  });
   
   function struct(props, name) {
   
     // DEBUG HINT: if the debugger stops here, the first argument is not a dict of types
     // mouse over the `props` variable to see what's wrong
-    assert(dict(Type).is(props), 'Invalid argument `props` supplied to `struct()`');
+    assert(dict(Str, Type).is(props), 'Invalid argument `props` supplied to `struct()`');
 
     // DEBUG HINT: if the debugger stops here, the second argument is not a string
     // mouse over the `name` variable to see what's wrong
@@ -23227,7 +23259,7 @@ module.exports = require('./lib/React');
         }
       }
   
-      if (!mut) { 
+      if (mut !== true) { 
         Object.freeze(this); 
       }
     }
@@ -23238,12 +23270,18 @@ module.exports = require('./lib/React');
       name: name
     };
   
-    Struct.is = function (x) { 
+    Struct.is = function isStruct(x) { 
       return x instanceof Struct; 
     };
   
-    Struct.update = update;
+    Struct.update = function updateStruct(instance, spec, value) {
+      return new Struct(update(instance, spec, value));
+    };
   
+    Struct.extend = function extendStruct(newProps, name) {
+      return struct([props].concat(newProps).reduce(mixin, {}), name);
+    };
+
     return Struct;
   }
 
@@ -23274,7 +23312,7 @@ module.exports = require('./lib/React');
       var type = Union.dispatch(value);
 
       // DEBUG HINT: if the debugger stops here, the `dispatch` static method returns no type
-      assert(isType(type), '%s.dispatch() returns no type', name);
+      assert(Type.is(type), '%s.dispatch() returns no type', name);
       
       // DEBUG HINT: if the debugger stops here, `value` can't be converted to `type`
       // mouse over the `value` and `type` variables to see what's wrong
@@ -23287,14 +23325,14 @@ module.exports = require('./lib/React');
       name: name
     };
   
-    Union.is = function (x) {
-      return types.some(function (type) {
+    Union.is = function isUnion(x) {
+      return types.some(function isType(type) {
         return type.is(x);
       });
     };
   
     // default dispatch implementation
-    Union.dispatch = function (x) {
+    Union.dispatch = function dispatch(x) {
       for (var i = 0, len = types.length ; i < len ; i++ ) {
         if (types[i].is(x)) {
           return types[i];
@@ -23308,7 +23346,7 @@ module.exports = require('./lib/React');
   function maybe(type, name) {
   
     // DEBUG HINT: if the debugger stops here, the first argument is not a type
-    assert(isType(type), 'Invalid argument `type` supplied to `maybe()`');
+    assert(Type.is(type), 'Invalid argument `type` supplied to `maybe()`');
   
     // makes the combinator idempotent
     if (getKind(type) === 'maybe') {
@@ -23337,7 +23375,7 @@ module.exports = require('./lib/React');
       name: name
     };
   
-    Maybe.is = function (x) {
+    Maybe.is = function isMaybe(x) {
       return Nil.is(x) || type.is(x);
     };
   
@@ -23377,14 +23415,14 @@ module.exports = require('./lib/React');
       name: name
     };
   
-    Enums.is = function (x) {
+    Enums.is = function isEnums(x) {
       return Str.is(x) && map.hasOwnProperty(x);
     };
   
     return Enums;
   }
   
-  enums.of = function (keys, name) {
+  enums.of = function enumsOf(keys, name) {
     keys = Str.is(keys) ? keys.split(' ') : keys;
     var value = {};
     keys.forEach(function (k) {
@@ -23400,9 +23438,6 @@ module.exports = require('./lib/React');
 
     var len = types.length;
 
-    // DEBUG HINT: if the debugger stops here, there are too few types (they must be at least two)
-    assert(len >= 2, 'Invalid argument `types` supplied to `tuple()`');
-
     // DEBUG HINT: if the debugger stops here, the second argument is not a string
     // mouse over the `name` variable to see what's wrong
     assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `tuple()`');
@@ -23411,9 +23446,6 @@ module.exports = require('./lib/React');
   
     function Tuple(value, mut) {
   
-      // DEBUG HINT: if the debugger stops here, you have used the `new` operator but it's forbidden
-      blockNew(this, Tuple);
-
       // DEBUG HINT: if the debugger stops here, the value is not one of the defined enums
       // mouse over the `value`, `name` and `len` variables to see what's wrong
       assert(Arr.is(value) && value.length === len, 'Invalid `%s` supplied to `%s`, expected an `Arr` of length `%s`', value, name, len);
@@ -23432,7 +23464,7 @@ module.exports = require('./lib/React');
         arr.push(expected(actual, mut));
       }
   
-      if (!mut) { 
+      if (mut !== true) { 
         Object.freeze(arr); 
       }
       return arr;
@@ -23441,6 +23473,7 @@ module.exports = require('./lib/React');
     Tuple.meta = {
       kind: 'tuple',
       types: types,
+      length: len,
       name: name
     };
   
@@ -23450,11 +23483,13 @@ module.exports = require('./lib/React');
       });
     };
   
-    Tuple.is = function (x) {
+    Tuple.is = function isTuple(x) {
       return Arr.is(x) && x.length === len && Tuple.isTuple(x);
     };
   
-    Tuple.update = update;
+    Tuple.update = function updateTuple(instance, spec, value) {
+      return Tuple(update(instance, spec, value));
+    };
   
     return Tuple;
   }
@@ -23462,7 +23497,7 @@ module.exports = require('./lib/React');
   function subtype(type, predicate, name) {
   
     // DEBUG HINT: if the debugger stops here, the first argument is not a type
-    assert(isType(type), 'Invalid argument `type` supplied to `subtype()`');
+    assert(Type.is(type), 'Invalid argument `type` supplied to `subtype()`');
     
     // DEBUG HINT: if the debugger stops here, the second argument is not a function
     assert(Func.is(predicate), 'Invalid argument `predicate` supplied to `subtype()`');
@@ -23498,17 +23533,21 @@ module.exports = require('./lib/React');
       name: name
     };
   
-    Subtype.is = function (x) {
+    Subtype.is = function isSubtype(x) {
       return type.is(x) && predicate(x);
     };
   
+    Subtype.update = function updateSubtype(instance, spec, value) {
+      return Subtype(update(instance, spec, value));
+    };
+
     return Subtype;
   }
 
   function list(type, name) {
   
     // DEBUG HINT: if the debugger stops here, the first argument is not a type
-    assert(isType(type), 'Invalid argument `type` supplied to `list()`');
+    assert(Type.is(type), 'Invalid argument `type` supplied to `list()`');
   
     // DEBUG HINT: if the debugger stops here, the third argument is not a string
     // mouse over the `name` variable to see what's wrong
@@ -23520,7 +23559,6 @@ module.exports = require('./lib/React');
     function List(value, mut) {
   
       // DEBUG HINT: if the debugger stops here, you have used the `new` operator but it's forbidden
-      blockNew(this, List);
 
       // DEBUG HINT: if the debugger stops here, the value is not one of the defined enums
       // mouse over the `value` and `name` variables to see what's wrong
@@ -23539,7 +23577,7 @@ module.exports = require('./lib/React');
         arr.push(type(actual, mut));
       }
   
-      if (!mut) { 
+      if (mut !== true) { 
         Object.freeze(arr); 
       }
       return arr;
@@ -23555,34 +23593,35 @@ module.exports = require('./lib/React');
       return x.every(type.is);
     };
   
-    List.is = function (x) {
+    List.is = function isList(x) {
       return Arr.is(x) && List.isList(x);
     };
   
-  
-    List.update = update;
+    List.update = function updateList(instance, spec, value) {
+      return List(update(instance, spec, value));
+    };
   
     return List;
   }
 
-  function dict(type, name) {
+  function dict(domain, codomain, name) {
   
     // DEBUG HINT: if the debugger stops here, the first argument is not a type
-    assert(isType(type), 'Invalid argument `type` supplied to `dict()`');
+    assert(Type.is(domain), 'Invalid argument `domain` supplied to `dict()`');
+
+    // DEBUG HINT: if the debugger stops here, the second argument is not a type
+    assert(Type.is(codomain), 'Invalid argument `codomain` supplied to `dict()`');
   
     // DEBUG HINT: if the debugger stops here, the third argument is not a string
     // mouse over the `name` variable to see what's wrong
     assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `dict()`');
 
     // DEBUG HINT: always give a name to a type, the debug will be easier
-    name = name || format('dict(%s)', getName(type));
+    name = name || format('dict(%s, %s)', getName(domain), getName(codomain));
 
     function Dict(value, mut) {
   
-      // DEBUG HINT: if the debugger stops here, you have used the `new` operator but it's forbidden
-      blockNew(this, Dict);
-
-      // DEBUG HINT: if the debugger stops here, the value is not one of the defined enums
+      // DEBUG HINT: if the debugger stops here, the value is not an object
       // mouse over the `value` and `name` variables to see what's wrong
       assert(Obj.is(value), 'Invalid `%s` supplied to `%s`, expected an `Obj`', value, name);
   
@@ -23594,14 +23633,17 @@ module.exports = require('./lib/React');
       var obj = {};
       for (var k in value) {
         if (value.hasOwnProperty(k)) {
+          // DEBUG HINT: if the debugger stops here, the `k` value supplied to the `domain` type is invalid
+          // mouse over the `k` and `domain` variables to see what's wrong
+          k = domain(k);
           var actual = value[k];
-          // DEBUG HINT: if the debugger stops here, the `actual` value supplied to the `type` type is invalid
-          // mouse over the `actual` and `type` variables to see what's wrong
-          obj[k] = type(actual, mut);
+          // DEBUG HINT: if the debugger stops here, the `actual` value supplied to the `codomain` type is invalid
+          // mouse over the `actual` and `codomain` variables to see what's wrong
+          obj[k] = codomain(actual, mut);
         }
       }
   
-      if (!mut) { 
+      if (mut !== true) { 
         Object.freeze(obj); 
       }
       return obj;
@@ -23609,122 +23651,137 @@ module.exports = require('./lib/React');
   
     Dict.meta = {
       kind: 'dict',
-      type: type,
+      domain: domain,
+      codomain: codomain,
       name: name
     };
   
     Dict.isDict = function (x) {
       for (var k in x) {
-        if (x.hasOwnProperty(k) && !type.is(x[k])) {
-          return false;
+        if (x.hasOwnProperty(k)) {
+          if (!domain.is(k) || !codomain.is(x[k])) { return false; }
         }
       }
       return true;
     };
   
-    Dict.is = function (x) {
+    Dict.is = function isDict(x) {
       return Obj.is(x) && Dict.isDict(x);
     };
   
   
-    Dict.update = update;
+    Dict.update = function updateDict(instance, spec, value) {
+      return Dict(update(instance, spec, value));
+    };
   
     return Dict;
   }
 
-  function func(Arguments, f, Return, name) {
-  
-    name = name || 'func()';
-    Arguments = Arr.is(Arguments) ? tuple(Arguments, 'Arguments') : Arguments;
+  function func(domain, codomain, name) {
 
-    // DEBUG HINT: if the debugger stops here, the first argument is not a type
-    assert(isType(Arguments), 'Invalid argument `Arguments` supplied to `func()`');
+    // handle handy syntax for unary functions
+    domain = Arr.is(domain) ? domain : [domain];
 
-    // DEBUG HINT: if the debugger stops here, the second argument is not a function
-    assert(Func.is(f), 'Invalid argument `f` supplied to `func()`');
+    // DEBUG HINT: if the debugger stops here, the first argument is not a list of types
+    assert(list(Type).is(domain), 'Invalid argument `domain` supplied to `func()`');
 
-    // DEBUG HINT: if the debugger stops here, the third argument is not a type (or Nil)
-    assert(Nil.is(Return) || isType(Return), 'Invalid argument `Return` supplied to `func()`');
-
-    // DEBUG HINT: if the debugger stops here, the third argument is not a string
-    // mouse over the `name` variable to see what's wrong
-    assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `func()`');
+    // DEBUG HINT: if the debugger stops here, the second argument is not a type
+    assert(Type.is(codomain), 'Invalid argument `codomain` supplied to `func()`');
 
     // DEBUG HINT: always give a name to a type, the debug will be easier
-    name = name || f.name || 'func';
-  
-    // makes the combinator idempotent
-    Return = Return || null;
-    if (isType(f) && f.meta.Arguments === Arguments && f.meta.Return === Return) {
-      return f;
-    }
-  
-    function fn() {
-  
-      var args = slice.call(arguments);
-  
-      // handle optional arguments
-      if (args.length < f.length) {
-        args.length = f.length; 
+    name = name || format('func([%s], %s)', domain.map(getName).join(', '), getName(codomain));
+
+    // cache the domain length
+    var domainLen = domain.length;
+
+    function Func(value) {
+
+      // automatically instrument the function if is not already instrumented
+      if (!func.is(value)) {
+        value = Func.of(value);
       }
-  
-      // DEBUG HINT: if the debugger stops here, the arguments of the function are invalid
-      // mouse over the `args` variable to see what's wrong
-      args = Arguments(args);
-  
-      /*jshint validthis: true */
-      var r = f.apply(this, args);
-  
-      if (Return) {
-        // DEBUG HINT: if the debugger stops here, the return value of the function is invalid
-        // mouse over the `r` variable to see what's wrong
-        r = Return(r);
-      }
-  
-      return r;
+
+      // DEBUG HINT: if the debugger stops here, the first argument is invalid
+      // mouse over the `value` and `name` variables to see what's wrong
+      assert(Func.is(value), 'Invalid `%s` supplied to `%s`', value, name);
+      
+      return value;
     }
-  
-    fn.is = function (x) { 
-      return x === fn; 
-    };
-  
-    fn.meta = {
+
+    Func.meta = {
       kind: 'func',
-      Arguments: Arguments,
-      f: f,
-      Return: Return,
+      domain: domain,
+      codomain: codomain,
       name: name
     };
-  
-    return fn;
-  }
 
-  function alias(type, name) {
-
-    // DEBUG HINT: if the debugger stops here, the first argument is not a type
-    assert(isType(type), 'Invalid argument `type` supplied to `alias()`');
-  
-    // DEBUG HINT: if the debugger stops here, the third argument is not a string
-    // mouse over the `name` variable to see what's wrong
-    assert(maybe(Str).is(name), 'Invalid argument `name` supplied to `alias()`');
-
-    // DEBUG HINT: always give a name to a type, the debug will be easier
-    name = name || 'alias(' + getName(type) + ')';
-
-    function Alias(value, mut) {
-      return type(value, mut);
-    }
-
-    Alias.is = function (x) {
-      return type.is(x);
+    Func.is = function isFunc(x) {
+      return func.is(x) && 
+        x.func.domain.length === domain.length && 
+        x.func.domain.every(function (type, i) {
+          return type === domain[i];
+        }) && 
+        x.func.codomain === codomain; 
     };
 
-    Alias.meta = type.meta;
-    Alias.name = name;
+    Func.of = function funcOf(f) {
 
-    return Alias;
+      // DEBUG HINT: if the debugger stops here, f is not a function
+      assert(typeof f === 'function');
+
+      // makes Func.of idempotent
+      if (Func.is(f)) {
+        return f;
+      }
+
+      function fn() {
+    
+        var args = slice.call(arguments);
+        var len = Math.min(args.length, domainLen);
+        
+        // DEBUG HINT: if the debugger stops here, you provided wrong arguments to the function
+        // mouse over the `args` variable to see what's wrong
+        args = tuple(domain.slice(0, len))(args);
+
+        if (len === domainLen) {
+
+          /* jshint validthis: true */
+          var r = f.apply(this, args);
+      
+          // DEBUG HINT: if the debugger stops here, the return value of the function is invalid
+          // mouse over the `r` variable to see what's wrong
+          r = codomain(r);
+      
+          return r;
+
+        } else {
+
+          var curried = Function.prototype.bind.apply(f, [this].concat(args));
+          var newdomain = func(domain.slice(len), codomain);
+          return newdomain.of(curried);
+
+        }
+    
+      }
+    
+      fn.func = {
+        domain: domain,
+        codomain: codomain,
+        f: f
+      };
+    
+      return fn;
+
+    };
+
+    return Func;
 
   }
+
+  // returns true if x is an instrumented function
+  func.is = function (f) {
+    return Func.is(f) && Obj.is(f.func);
+  };
 
   return {
 
@@ -23732,11 +23789,11 @@ module.exports = require('./lib/React');
       mixin: mixin,
       merge: merge,
       format: format,
-      isType: isType,
       getName: getName,
       getKind: getKind,
-      isKind: isKind,
-      slice: slice
+      slice: slice,
+      shallowCopy: shallowCopy,
+      update: update
     },
 
     options: options,
@@ -23765,8 +23822,7 @@ module.exports = require('./lib/React');
     subtype: subtype,
     list: list,
     dict: dict,
-    func: func,
-    alias: alias
+    func: func
   };
 }));
 
@@ -23795,6 +23851,7 @@ $(function () {
   var createForm = t.form.createForm;
   var createList = t.form.createList;
   var radio = t.form.radio;
+  var select = t.form.select;
 
   //
   // setup
@@ -23826,7 +23883,8 @@ $(function () {
     {id: 'nestedLists', label: '13. Nested lists'},
     {id: 'goodies', label: '14. Bootstrap goodies'},
     {id: 'horizontal', label: '15. Horizontal forms'},
-    {id: 'customInput', label: '16. Custom input'}
+    {id: 'customInput', label: '16. Custom input'},
+    {id: 'multiple', label: '17. Multiple select'}
   ];
 
   var examples = {};